fix a bug

+++ model/utils/loss.py
@@ -46,7 +46,7 @@ def delta_loss(pred_delta, target_delta, anchor_label, sigma):
     loss = _smooth_l1_loss(pred_delta, target_delta, weight, sigma)
     
     # ignore gt_label==-1 for rpn_loss
-    loss = loss / (anchor_label.data >=0).sum() 
+    loss = loss / (anchor_label.data >=0).sum().type_as(loss) 
     return loss
 
 if __name__ == '__main__':
